Move `DefaultAudioSink.AudioProcessorChain` to `AudioProcessorChain`
Split inner interface into separate file, which will go in common
module. The old interface will be deprecated and extends the new.

#cleanup

PiperOrigin-RevId: 483732226
